+++ R/beta_temporal.R
@@ -11,7 +11,7 @@
 #' a calling function.
 #'
 #' @return A vector with beta results (total, replacement,
-#' and richness differences).
+#' richness difference, and ratio).
 #'
 temp.beta.vec <- function(x, nspp, spp, tree, resu, ...) {
   # Check if 'x' contains only NA values and return NA values for
@@ -47,9 +47,11 @@ temp.beta.vec <- function(x, nspp, spp, tree, resu, ...) {
 #' @param bin1 A SpatRaster with presence-absence data (0 or 1)
 #' for a set of species.
 #' @param bin2 A SpatRaster with presence-absence data (0 or 1)
-#' for a set of species.
-#' @param tree It can be a data frame with species traits or a
-#' phylogenetic tree.
+#' for a set of species. Species names in 'bin2' and 'bin1' must
+#' match!
+#' @param tree It can be a 'data.frame' with species traits or a
+#' 'phylo' with a rooted phylogenetic tree. Species names in 'tree',
+#' 'bin1', and 'bin2' must match!
 #' @param filename Character. Save results if a name is provided.
 #' @param cores A positive integer. If cores > 1, a 'parallel'
 #' package cluster with that many cores is created and used.
@@ -74,8 +76,13 @@ temp.beta.vec <- function(x, nspp, spp, tree, resu, ...) {
 #' and methodological framework for exploring and explaining
 #' pattern in presence - absence data. - Oikos 120: 1625–1638.
 #'
+#' @references Hidasi-Neto, J. et al. 2019. Climate change will
+#' drive mammal species loss and biotic homogenization in the
+#' Cerrado Biodiversity Hotspot. - Perspectives in Ecology and
+#' Conservation 17: 57–63.
+#'
 #' @return A SpatRaster with beta results (total, replacement,
-#' and richness differences).
+#' richness difference, and ratio).
 #' @export
 #'
 #' @examples
@@ -98,28 +105,8 @@ temp.beta <- function(bin1,
                       tree,
                       filename = "",
                       cores = 1, ...) {
-  # Check if 'bin1' and 'bin2' are NULL or invalid (not SpatRaster)
-  if (is.null(bin1) || !inherits(bin1, "SpatRaster")) {
-    stop("'bin1' must be a SpatRaster.")
-  }
-  if (is.null(bin2) || !inherits(bin2, "SpatRaster")) {
-    stop("'bin2' must be a SpatRaster.")
-  }
-
-  # Check if coordinates of both rasters are geographic
-  if (!terra::is.lonlat(bin1) | !terra::is.lonlat(bin2)) {
-    stop("Both rasters must have geographic coordinates.")
-  }
-
-  # Check if both rasters have at least 2 layers
-  if (terra::nlyr(bin1) < 2 | terra::nlyr(bin2) < 2) {
-    stop("Both rasters must have at least 2 layers.")
-  }
-
-  # Check if the names of bin1 and bin2 match
-  if (!identical(names(bin1), names(bin2))) {
-    stop("The names of the rasters do not match.")
-  }
+  # Initial tests
+  inputs_chk(bin1 = bin1, bin2 = bin2, tree = tree)
 
   # Get number of species
   nspp <- terra::nlyr(bin1)
@@ -128,7 +115,7 @@ temp.beta <- function(bin1,
   spp <- names(bin1)
 
   # Create a numeric vector to store results for Btotal, Brepl,
-  # and Brich
+  # Brich
   resu <- numeric(3)
 
   # Apply the function to the SpatRaster objects 'bin1' and 'bin2'
@@ -140,11 +127,7 @@ temp.beta <- function(bin1,
                       spp = spp,
                       cores = cores, ...)
   } else {
-    # Check if 'tree' object is valid (either a data.frame or a
-    # phylo object)
-    if (!inherits(tree, c("data.frame", "phylo"))) {
-      stop("'tree' must be a data.frame or a phylo object.")
-    }
+
     res <- terra::app(c(bin1, bin2),
                       temp.beta.vec,
                       resu = resu,
@@ -153,9 +136,12 @@ temp.beta <- function(bin1,
                       spp = spp,
                       cores = cores, ...)
   }
+  # Calculate beta ratio (Brepl / Btotal) and store it
+  # See Hidasi-Neto et al. (2019)
+  res <- c(res, res[[2]]/res[[1]])
 
   # Define names for the output based on the type of 'tree'
-  lyrnames <- c("Btotal", "Brepl", "Brich")
+  lyrnames <- c("Btotal", "Brepl", "Brich", "Bratio")
   if (missing(tree)) {
     names(res) <- paste0(lyrnames, "_TD")
   } else if (inherits(tree, "data.frame")) {
@@ -171,6 +157,6 @@ temp.beta <- function(bin1,
                        overwrite = TRUE, ...)
   }
 
-  # Return the numeric vector with beta diversity values
+  # Return beta diversity values
   return(res)
 }